refactor: 음식점 정보, 자주 가는 음식점 버튼, 카테고리 이미지 컴포넌트 분리
