+++ src/app/component/FavoriteButton/FavoriteButton.css
@@ -1,0 +1,15 @@
+.restaurant__favorite-button {
+  position: absolute;
+  top: 16px;
+  right: 8px;
+
+  width: 32px;
+  height: 32px;
+
+  background: transparent;
+  background-image: url('../../../asset/image/favorite-icon-lined.png');
+}
+
+.restaurant__favorite-button.favorited {
+  background-image: url('../../../asset/image/favorite-icon-filled.png');
+}
